+++ Sources/BraintreeCore/Public/BraintreeCore/BraintreeCore.h
@@ -20,7 +20,6 @@ FOUNDATION_EXPORT const unsigned char BraintreeCoreVersionString[];
 #import <Braintree/BTLogger.h>
 #import <Braintree/BTPostalAddress.h>
 #import <Braintree/BTPaymentMethodNonce.h>
-#import <Braintree/BTPayPalIDToken.h>
 #import <Braintree/BTPaymentMethodNonce.h>
 #import <Braintree/BTViewControllerPresentingDelegate.h>
 #import <Braintree/BTPreferredPaymentMethods.h>
@@ -40,7 +39,6 @@ FOUNDATION_EXPORT const unsigned char BraintreeCoreVersionString[];
 #import <BraintreeCore/BTLogger.h>
 #import <BraintreeCore/BTPostalAddress.h>
 #import <BraintreeCore/BTPaymentMethodNonce.h>
-#import <BraintreeCore/BTPayPalIDToken.h>
 #import <BraintreeCore/BTPaymentMethodNonce.h>
 #import <BraintreeCore/BTViewControllerPresentingDelegate.h>
 #import <BraintreeCore/BTPreferredPaymentMethods.h>