+++ Sources/BraintreeVenmo/BTVenmoDriver.m
@@ -35,6 +35,7 @@
 
 NSString * const BTVenmoDriverErrorDomain = @"com.braintreepayments.BTVenmoDriverErrorDomain";
 NSString * const BTVenmoAppStoreUrl = @"https://itunes.apple.com/us/app/venmo-send-receive-money/id351727428";
+NSInteger const NetworkConnectionLostCode = -1005;
 
 @implementation BTVenmoDriver
 
@@ -51,7 +52,7 @@ static BTVenmoDriver *appSwitchedDriver;
 
 - (instancetype)initWithAPIClient:(BTAPIClient *)apiClient {
     if (self = [super init]) {
-        _apiClient = [apiClient copyWithSource:BTClientMetadataSourceVenmoApp integration:apiClient.metadata.integration];
+        _apiClient = apiClient;
     }
     return self;
 }
@@ -125,19 +126,19 @@ static BTVenmoDriver *appSwitchedDriver;
             completionBlock(nil, configurationError);
             return;
         }
-
+        
         NSError *error;
         if (![self verifyAppSwitchWithConfiguration:configuration error:&error]) {
             completionBlock(nil, error);
             return;
         }
-
+        
         NSString *merchantProfileID = venmoRequest.profileID ?: configuration.venmoMerchantID;
         NSString *bundleDisplayName = [self.bundle objectForInfoDictionaryKey:@"CFBundleDisplayName"];
-
+        
         BTMutableClientMetadata *metadata = [self.apiClient.metadata mutableCopy];
         metadata.source = BTClientMetadataSourceVenmoApp;
-
+        
         NSMutableDictionary *inputParams = [@{
             @"paymentMethodUsage": venmoRequest.paymentMethodUsageAsString,
             @"merchantProfileId": merchantProfileID,
@@ -158,6 +159,9 @@ static BTVenmoDriver *appSwitchedDriver;
         
         [self.apiClient POST:@"" parameters:params httpType:BTAPIClientHTTPTypeGraphQLAPI completion:^(BTJSON *body, __unused NSHTTPURLResponse *response, NSError *err) {
             if (err) {
+                if (err.code == NetworkConnectionLostCode) {
+                    [self.apiClient sendAnalyticsEvent:@"ios.pay-with-venmo.network-connection.failure"];
+                }
                 NSError *error = [NSError errorWithDomain:BTVenmoDriverErrorDomain
                                                      code:BTVenmoDriverErrorTypeInvalidRequestURL
                                                  userInfo:@{NSLocalizedDescriptionKey: @"Failed to fetch a Venmo paymentContextID while constructing the requestURL."}];
@@ -173,7 +177,7 @@ static BTVenmoDriver *appSwitchedDriver;
                 completionBlock(nil, error);
                 return;
             }
-
+            
             NSURL *appSwitchURL = [BTVenmoAppSwitchRequestURL appSwitchURLForMerchantID:merchantProfileID
                                                                             accessToken:configuration.venmoAccessToken
                                                                         returnURLScheme:self.returnURLScheme
@@ -181,6 +185,7 @@ static BTVenmoDriver *appSwitchedDriver;
                                                                             environment:configuration.venmoEnvironment
                                                                        paymentContextID:paymentContextID
                                                                                metadata:self.apiClient.metadata];
+            
             [self performAppSwitch:appSwitchURL shouldVault:venmoRequest.vault completion:completionBlock];
         }];
     }];
@@ -198,6 +203,9 @@ static BTVenmoDriver *appSwitchedDriver;
               parameters:params
               completion:^(BTJSON *body, __unused NSHTTPURLResponse *response, NSError *error) {
                   if (error) {
+                      if (error.code == NetworkConnectionLostCode) {
+                          [self.apiClient sendAnalyticsEvent:@"ios.pay-with-venmo.network-connection.failure"];
+                      }
                       [self.apiClient sendAnalyticsEvent:@"ios.pay-with-venmo.vault.failure"];
                       self.appSwitchCompletionBlock(nil, error);
                   } else {
@@ -268,6 +276,9 @@ static BTVenmoDriver *appSwitchedDriver;
 
             [self.apiClient POST:@"" parameters:params httpType:BTAPIClientHTTPTypeGraphQLAPI completion:^(BTJSON *body, __unused NSHTTPURLResponse *response, NSError *error) {
                 if (error) {
+                    if (error.code == NetworkConnectionLostCode) {
+                        [self.apiClient sendAnalyticsEvent:@"ios.pay-with-venmo.network-connection.failure"];
+                    }
                     [self.apiClient sendAnalyticsEvent:@"ios.pay-with-venmo.appswitch.handle.client-failure"];
                     self.appSwitchCompletionBlock(nil, error);
                     self.appSwitchCompletionBlock = nil;